debug: move colors to debugColors.ts from debugToolBar
